backend/requirements.txt
# ── Framework web ─────────────────────────────────────────────────────────────
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.1
werkzeug==3.0.3

# ── Base de données SQL Server ─────────────────────────────────────────────────
pyodbc==5.1.0
pymssql==2.3.1
SQLAlchemy==2.0.31

# ── Authentification ───────────────────────────────────────────────────────────
PyJWT==2.8.0

# ── Environnement ──────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Data science / ML (ajoutés pour la prédiction Chl-a) ──────────────────────
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0     # utilitaires ML (preprocessing, métriques)
catboost>=1.2.0         # chargement du modèle .cbm natif CatBoost
openpyxl>=3.1.0         # lecture fichiers .xlsx
xlrd>=2.0.1             # lecture fichiers .xls (legacy)
